// Project/dff_library/ui_components/components/DataTables/utility/enum.ts
export enum DataTableTypes {
  DataTableFullScreen = 'DataTableFullScreen',
  DataTableWithMoreRows = 'DataTableWithMoreRows',
  TitleOnBothSidesHorizontalScroll = 'TitleOnBothSidesHorizontalScroll',
  TitleOnBothSidesVerticalScroll = 'TitleOnBothSidesVerticalScroll',
  TitleOnLeft = 'TitleOnLeft',
}